Moved item file into classes folder
